* fixed lack of connection in Cellular mode (migrated to OkHTTP which works stable here) * fixed some OpenWeatherModel parameters type mismatch * fixed calling from non-ui thread crash in details

+++ app/src/main/java/org/pdnk/goodweather/Provider/OpenWeatherModel/Main.java
@@ -10,7 +10,7 @@ public class Main {
     public double temp;
     @SerializedName("pressure")
     @Expose
-    public int pressure;
+    public double pressure;
     @SerializedName("humidity")
     @Expose
     public int humidity;
