+++ app/src/main/java/org/pdnk/goodweather/Provider/OpenWeatherModel/Wind.java
@@ -11,6 +11,6 @@ public class Wind {
     public double speed;
     @SerializedName("deg")
     @Expose
-    public int deg;
+    public double deg;
 
 }
